# src/main/resources/api/service_openapi.yaml
openapi: 3.0.0
info:
  title: Cloud Resource Janitor
  description: |
    Janitor service to cleanup cloud resources created by Terra CRL.
  version: 0.0.0

# This file has extra dependencies from: https://github.com/DataBiosphere/terra-cloud-resource-lib/blob/master/cloud-resource-schema/src/main/resources/cloud_resources_uid.yaml
servers:
  - url: /
security:
  - googleoauth:
      - openid
      - email
      - profile
  - bearerAuth: []
paths:
  '/status':
    get:
      security: []
      description: |
        Returns the operational status of the service
      operationId: serviceStatus
      tags:
        - unauthenticated
      responses:
        200:
          description: Service is functional
          $ref: '#/components/responses/StatusResponse'
        500:
          description: Service is broken
          $ref: '#/components/responses/StatusResponse'

  # TODO(PF-57) Add a shutdown to only be callable by Kubernetes in production and to be hooked to Pod termination.
...
components:
...
  schemas:
...
    SystemStatus:
      type: object
      properties:
        ok:
          type: boolean
          description: status of this service
        systems:
          type: object
          additionalProperties:
            type: object
            properties:
              ok:
                type: boolean
              messages:
                type: array
                items:
                  type: string
...
  responses:
...
    StatusResponse:
      description: common status response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SystemStatus'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    authorization:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          scopes:
            openid: open id authorization
            email: email authorization
            profile: profile authorization